Split version funcs out of bundle-gen
In preparation for using them to compute the version string embedded in
go builds, this commit extracts version-related functions and related
consts out of bundle-gen into a separate hack/version.py library. A main
routine is added so the file can be invoked standalone to produce the
same version string bundle-gen would use for the OperatorHub bundles:

```
[efried@efried hive]$ ./hack/version.py 2>/dev/null
1.2.3604-aebcdb3
```

A subsequent commit will plumb this into the Makefile.

HIVE-1717