machinepool: remove finalizer for fake CD
The "normal" machinepool controller flow does all the things, and then
at the end checks whether the machinepool is deleted. If so, it removes
the finalizer and bails.

For fake clusters, we are bailing before "all the things". But we were
not deleting the finalizer in that case.

To make fake clusters behave more like real clusters, this commit adds
finalizer removal to the fake cluster flow.