Replace cluster-api-provider-aws with machineapi.
APIs for AWSMachineProviderConfig have moved from
cluster-api-provider-aws to the machineapi. We can no longer use the
universal decoder because the AWSMachineProviderConfig is not a
registered type within the machineapi. Other providers (GCP & Azure)
have installer adjacent libs that provide a method for
ProviderSpecFromRawExtension but I was not able to find one for AWS so
the universal decoder has been replaced with json.Unmarshal within
decodeAWSMachineProviderSpec.