Azure changes from vendoring installer to bdbb3b1.
Installer Azure MachineSet generation adds an additional argument for
HyperVGeneration which is based on instance disk type and divined via
the Azure client.

The installer now creates a HyperVGen2 image in addition to a
HyperVGen1 image and uses the determined HyperVGen to decide which
image to use for the MachineSet. To ensure that Hive can create
MachineSets for older versions of OpenShift we must check that a
HyperVGen2 image exists before trying to use it.

Update Azure OSDisk type to include DiskType and DiskEncryptionSet
fields and plumb those fields into machinepool -> machineset
generation.