Further filter ODF security group filter by VPC ID
Security group names may be the same within an AWS account (belonging
to different VPCs) which means filtering on the name alone is
insufficient. The VPC ID is now used to further filter the correct
security group when configuring the ODF security group in worker
machineset manifests. The VPC ID is discovered by using the AWS client
to query subnets from the worker MachinePool. BYO VPC is required when
using ODF so it is expected that the worker MachinePool platform will
contain subnets we can use to discover the VPC ID.