Add replica ordinal to clustersync status
In an environment with multiple hive-clustersync replicas and a lot of
clustersync activity, a barrier to debugging was figuring out which
hive-clustersync replica was responsible for a given clusterdeployment's
syncsets. The usual way would be to grep the logs for the CD name, which
can take a nontrivial amount of time across multiple logs.

This commit adds a field to ClusterSync.Status called
ControlledByReplica. It is a pointer (so it doesn't default to zero) to
an integer indicating the ordinal of the replica reconciling it.