clustersync: fix missing memcpy
71c1839b / #1572 fixed an ordering problem in the clustersync controller
to address "moving" a resource from one syncset to another. In so doing,
it introduced a temporary list to look for the deletions. However, the
list was being aliased rather than copied, and the result was that
scenarios with multiple syncsets could end up ignoring changes to a
subset of them. Fix.

HIVE-1891