Set HOME to a writable dir for CI
In the CI environment, `$HOME` is set to `/` (the root directory), but
that directory is not writable by the test process, so things like pip
install (even with `--user`) fail because they want to create things
like `/.local`.

With this commit, our Makefile detects when `$HOME==/` and overrides it
to `/tmp/home`, which should be writable in the CI environment.

HIVE-1717