hibernation controller: Smooth a couple of state transitions
- If syncsets have not yet been applied, we want to delay requeue to
avoid thrashing while we give that time to happen. However, the check
for the syncset state was brittle, relying on a status condition. This
commit changes the controller to check for that state idempotently and
plumb that value through to where that delay is computed. This will
result in faster resume times (avoiding spurious delays) for non-new
clusters.
- If the hibernation controller restarts (or reactivates after having
been paused) at a particular time, it's possible for machines to be
started, but the remote API server isn't yet up; and this used to result
in PowerState incorrectly hanging at `StartingMachines`. Fix it to be
`WaitingForNodes` instead.