Remove NoMatchingRoute53Zone regex
This regex hasn't been seen in recent memory, and should be "impossible"
based on the current state of the code. Remove it for now. If the error
manifests again, it'll show up as "Unknown" and be routed through the
usual triage procedures, the result of which may be one or more of:
- Fixing the root cause in code
- Reinstating the regex
- Adding it to RetryReasons
- Nothing

HIVE-1933