Ensure correct time reported for transition metrics
Fetching cluster deployment conditions returns a pointer, and that
pointer is subject to change as other controllers update the CR, and
those conditions get sorted into a different order.

A bug was caught while testing powerstate transitioning metrics, where,
by the time we reported the hibernation transition metric, the ready
condition was pointing to some other condition of cluster deployment.
This commit ensures that before reporting the metric, we re-fetch the
relevant condition