Don't rely on WaitingForNodes
While checking for cluster operators, we always assume we would have
encountered WaitingForNodes, however that is not necessary as we could
have requeued for some other reason after machines were running, and by
the time we check, the nodes could have already been ready.
Allow for the clusteroperator check to proceed for other states as well,
like StartingMachines or WaitingForMachines.

- Also gotten rid of the redundant duration conversions