WIP: Clone the hibernation controller => powerstate
In anticipation of rewriting the thing, clone the hibernation controller
to a new package/controller called powerstate.

Choose which controller to use via HiveConfig.Spec.DisabledControllers
as follows:
- Signal that you want to use the powerstate controller by adding
`hibernation` to the list (but omitting `powerstate`).
- Otherwise, if `hibernation` is absent, we will disable the
powerstate controller (whether `powerstate` is present or not). This
preserves the existing default behavior.
- Both controllers can be disabled by specifying both names.

TODO: card