Refactor: Include ClusterProvision in InstallManager
In preparation for needing to plumb some data from the ClusterProvision
into the invocations of the openshift-install command, add a new field
to InstallManager containing the actual ClusterProvision object. This
allows us to change a bunch of function prototypes to remove it as an
argument, as those functions can get it from the InstallManager argument
they already accept (or are methods on).

HIVE-1966