Fix clearing of CS Failing metric
ClusterSync Failing metric must be cleared when clustersync is no longer
failing, however the code path that evoked clearing of that metric in
clustersync controller when the failing condition was changed to success
was somehow never evoked.
- This commit moves the clearing logic under this criterion to the
  metrics controller. The controller assumes that if any clustersync is
not failing than the corresponding metric should be cleared.
- Since the code path in clustersync controller wasn't being evoked,
  this commit also reverts to the previous code, and moves the clearing
logic to the metrics controller, thereby ensuring we are only setting or
clearing the metric in either the metrics controller or the
clusterdeployment controller