hiveutil: supply/consume additional log fields via env
- The `hiveutil` executable is changed to look for and process an
environment variable, `HIVE_ADDITIONAL_LOG_FIELDS`. This a JSON string
representing a flat map of key/value pairs to be added to all log lines
produced by `hiveutil` (and passed through to the installer for the same
purpose).
- Each instance where hive code invokes `hiveutil` via `Job`s is wired
to copy such JSON from the `hive.openshift.io/additional-log-fields`
annotation on the object on which it is operating, to the env of the
`Job` that will invoke `hiveutil`.

HIVE-1966