Make ProvisionStopped terminal (f'real this time)
ProvisionStopped is supposed to be terminal. That is, once you're in
that state, you should never leave it.

In certain circumstances, like reconciling DNSZones under managed DNS,
it was possible to reach ProvisionStopped without ever having created a
ClusterProvision. The logic that decides whether to "retry" provisioning
would happily do so in this case. The result would be that the
ProvisionStopped condition would be cleared (possibly to return later).

This commit adds an early check for the ProvisionStopped condition and
refuses to proceed if it's True.

HIVE-1881