ScaleMode: Mount data plane kubeconfig secret
In scale mode, the operator now checks for the existence of the data
plane kubeconfig secret in the target namespace and projects it onto the
hive-controllers, hive-clustersync, and hiveadmission pods as a
volume+mount. An environment variable is added to those pods indicating
the path to the mounted kubeconfig file. In a subsequent commit, the
containers will key off of the existence of that environment variable to
know they need to load the kubeconfig and use it for data plane
operations.

HIVE-1862