ScaleMode: Split Control/Data Plane clients in controllers
In scale mode, each controller needs to understand that it may need to
be talking to the data plane and/or the control plane for each client
call. Building on earlier commits, we trigger off the
`DATA_PLANE_KUBECONFIG` environment variable to determine whether the
split is necessary.

Since most controllers only need to talk to the data plane, we make that
the default client in the reconcilers. If the `DATA_PLANE_KUBECONFIG`
variable is present, we build this client based on the kubeconfig in
that secret; otherwise we build it from the manager's REST config as
usual.

Controllers that need to perform control plane operations -- e.g. to
provision/deprovision clusters -- get an additional client in their
reconciler object. In either mode this is built from the manager's REST
config. I.e. for these controllers, in normal (non scale) mode, both
clients are identical.

One more tweak was necessary: the clusterpool controller creates a
namespace for each ClusterDeployment in a pool; and it and the
clusterclaim controller copy RBAC into that namespace. In scale mode,
this namespace and its RBAC must be created in both the data plane (to
store the CD, secrets, etc.) and the control plane (to run the
installer).

HIVE-1862