Tech Debt: Remove install log PVC functions
We stopped using a PVC to store install logs via #1189 / cb5142fd. These
functions were left around to clean up PVCs on installs that had used
this mechanism previously. It has been ~20mo, which should be long
enough for that to have happened in any reasonable deployment. Resolve
the TODOs and clean up the functions.