Tech Debt: Remove hibernation conditions for OCP <=4.4.8
OpenShift 4.4 has been out of support for a year and a half now [1], so
it should be safe to remove these version checks from the hibernation
controller.

Existing unit tests around "hibernation unsupported" are altered to use
missing actuator instead of OCP version. In working through this, a bug
was discovered: Unsupported CDs should be assumed Running as soon as
they're Installed. HIVE-2016 was opened and a comment left to fix this.

[1] https://access.redhat.com/support/policy/updates/openshift