Move secret/configmap retrieval to hiveutil for prov/deprov
Pods created by hive-controllers require information from secrets and
configmaps (for example, cloud credentials) in order to function.
Previously these were passed into the pods from the hive-controllers side via
environment variables and volumes. In scale mode (see HIVE-1861) we're
going to need that same business logic -- which parts of which objects
are needed in what form -- on the hiveutil side.

In order to avoid duplication of that logic, we're moving it *all* to
the hiveutil side, even for non-scale mode.

The cloud credentials secret and other secrets/configmaps are no longer
passed through via the pod spec. Instead, the hiveutil command itself
uses a local client to load those secrets and set up the same
environment variables and files that were previously passed through on
the pod.

HIVE-2021