Make install log regexes case-insensitive
Making programmatic use of strings intended for human consumption is a
fragile endeavor. These strings are generally not going to be considered
part of an API, and are therefore subject to change upstream without
warning or versioning.

Such a change was noticed recently: the first letter of an error string
used to be uppercase and is now lowercase. Hypothesis: someone saw the
yellow squiggly line in their IDE that said error messages shouldn't
start with uppercase letters, and "fixed" it. Lest such fixes should
prove prevalent, this commit changes our install log regex matching to
be case-insensitive across the board.

The theoretical risk is that we will now match a string incorrectly.
This is considered to be negligible.

HIVE-2018