Add install success/fail metrics
To come up with SLIs indicating the amount of time taken for a
successful cluster install, metrics that report on how long it took for
a cluster install to succeed or fail would be really helpful.

This commit adds hive_cluster_deployment_install_failure_total and
hive_cluster_deployment_install_success_total metrics, which also report
on the cluster_type, cloud platform, region, cluster version, count of
replicas/workers and the number of install attempt pertaining to the
provision attempt.