Assume PowerState "Running" for hibernation-unsupported clouds
Platforms that don't have a hibernation actuator...

- BareMetal
- OpenStack
- VSphere
- Ovirt
- AgentBareMetal
- None

...used to show up with their PowerState empty. With this commit, we
assume they are running, since we can't actually tell. It's possible
this is a lie -- e.g. if the instances get shut down from the cloud
provider -- but a) this is a better UX than empty; and b) it makes other
bits of the code operate correctly.

HIVE-2016