Partially unwind poolAlwaysRunning hack
In a previous commit, we started to spoof the PowerState of
hibernation-unsupported clusters as Running once they're Installed. This
means we no longer need to impose a special check to place such clusters
in the Assignable bucket for ClusterPools.

We still keep the artificial check to ensure that runningCount==size for
these platforms though. This is to force the user to acknowledge that
such clusters will never be hibernated.

HIVE-2016