Add support to override effective in-cluster credentials mode
ARO requires different set of credentials for installer and in-cluster
operations. This commit adds an annotation that would override
the existing credentials mode in the install-config. The reason we
do this hack is because in ARO we do not want to pass installer
credentials inside the cluster. This is done by setting credentials
mode to `Manual`. We want to instead pass the credentials supplied
by manifestSecretRef and use the same by overriding credentials
mode from `Manual` to `Passthrough`.