ScaleMode: Switch Job ownership to labels/finalizers
Previously we used OwnerReferences to indicate that Jobs we create are
owned by ClusterDeployment (for the imageset job), ClusterProvision
(installer), or ClusterDeprovision (uninstaller). These were used for
two purposes:
- Watch(Job, EnqueueRequestForOwner)
- Automatic/implicit deletion of the Job when the owning object was
deleted.

In scale mode, Jobs are in the control plane and the other objects are
in the data plane; so we can't use owner references for those cases.

This commit gets rid of those owner references and adds code to replace
their functionality:
- We now Watch(Job, EnqueueRequestFromMapFunc) with a map func that uses
labels on the Job (the labels were already being set) to determine the
owner.
- The controllers for these objects get new logic to delete the owned
Jobs -- again identified by their labels -- and use a new finalizer to
block deletion of the parent object until that's done.

HIVE-1862