Refine error checks
Currently, the checks for BootstrapFailed are only checking for
the generic error message. The installer checks to see if the
API came up and if the status is set to complete
in the bootstrap config map before exiting with BootstrapFailed
error and this can be checked to possibly see if the API came up
before it errored out.