Allow optional cluster-type labels
Allow users to add more cluster-type labels for metrics logging. These
labels should be applied to all ClusterDeployment and ClusterProvision
metrics that are logged per cluster-type. Optional labels can be
provided via metricsConfig section in the hiveConfig. Users can name the
label on the metric and assign it to a corresponding clusterDeployment
(or provision) label.

- While working on these changes, 2 metrics from ClusterProvision
  controller (metricClusterProvisionsTotal and metricInstallErrors) were
  moved to clusterProvision\metrics to collate all the relevant metrics
  and support was added for optional labels.
- Registering of metrics was moved later instead of init() time because
  the latter is executed before we could read the hiveConfig.
  Consequently the metrics needed to be explicitly registered for unit
  test purposes to avoid nil panics