Relocate: Ignore kube-controller-manager .crt ConfigMaps
Starting in k8s 1.21, the kube-controller-manager (and presumably
OpenShift's fork thereof) creates/reconciles `kube-root-ca.crt` and
`openshift-service-ca.crt` ConfigMaps in every namespace. On relocate,
as we're copying objects, we'll try to copy these from the source to the
destination cluster, fail because the object (with the same name)
already exists, try to delete it, and then race with the kcm trying to
create it again. If the relocate controller loses the race, it'll fail
the relocate and try again. It might succeed eventually if it wins the
race, but in at least one case (see associated issue/card) it never won.

With this commit, we ignore these two ConfigMaps, based on their names
(hardcoded).

HIVE-2080

closes #1932