Name inventory test cluster to be caught by lambda
1eb917189 / #1918 tried to rename the cluster we generate for testing
ClusterPool inventory so it would be caught by our leak detector lambda.
We erroneously thought all we had to do was prefix the name with
`hiveci-` when in fact the whole name gets parsed to glean the timestamp
(when the test started) and the PR number under test. As such, when one
of these clusters leaked, it was causing an exception in the lambda
function like:

```
Traceback (most recent call last):
  File "/var/task/ci_monitor_lambda_function.py", line 58, in lambda_handler
    formatted, count = build_report_text(vpcs, debug=debug)
  File "/var/task/ci_monitor_lambda_function.py", line 84, in build_report_text
    parsed = parse_vpc(vpc)
  File "/var/task/ci_monitor_lambda_function.py", line 23, in parse_vpc
    ret['pr'] = int(hex_pr, 16)
```

Due to length constraints, we can't use a suffix; so instead we're going
to update the lambda to recognize an additional prefix, `cdcci-` and
keep the rest of the cluster name the same.