HiveConfig.Spec.DeploymentConfig: customize controller resources
Heretofore, container resource requests for the deployments/statefulSets
managed by hive-operator were hardcoded into the operator's bindata,
with no way to customize them.

This commit adds a new field to HiveConfig.Spec called DeploymentConfig.
In it, you may list a subsection for each of the
deployments/statefulSets (hive-controllers, hive-clustersync, and/or
hiveadmission) containing a `resources`, which is a standard
corev1.ResourceRequirements. For example:

```yaml
  deploymentConfig:
  - deploymentName: hive-controllers
    resources:
      requests:
        memory: 256Mi
  - deploymentName: hive-clustersync
    resources:
      requests:
        cpu: 30m
        memory: 257Mi
```

hive-operator will redeploy the controllers (per usual when HiveConfig
is edited), copying the specified `resources` into the container spec
for the named deployment/statefulSet.

Note that this is separate from, and unrelated to,
HiveConfig.Spec.ControllersConfig, which tailors the behavior of
specific controllers running within the containers.

HIVE-2091