Apply Opts pattern to dynamic labels
This commit implements some of the changes recommended based on PR
comments. Main change is applying Opts pattern to the interface for
metrics with dynamic labels and apply a buildLabels helper to ensure
there is no overlap between fixed and optional labels as we define and
observe the corresponding metrics.

A side effect to the buildLabels helper is that we no longer need to
build a curried vector for a metric before observing it - we instead
build the map for labels to value and observe them via `With` (or
`GetMetricWith`) method, which is similar to how we observe metrics
without optional label support.