modcheck: handle multiple replacements for same lib
We resolved dependabot issues (see previous commit) by using a new-to-us
syntax for the `replace` directive in go.mod, where the left-hand side
of the arrow -- the version being replaced -- is specified at a
particular version. Our `modcheck` utility was not set up to handle this
properly: it was only using the library path as a map key for the
structures used to compare the two go.mod files. Fortunately, this
resulted in spurious failures; if the last replacement for a given
library happened to match between the two files, we wouldn't have
noticed.

This commit resolves the issue by including the replaced library's
version, if specified, in the map key used to compare the two go.mod
files. Thus all replacements for a given library are present in the maps
when we compare them, rather than just the last one.