Added hiveconfig.awsPrivateLink.SharedHostedZoneDomains
Added a new field to enable sharing of AWS dns zones for aws privatelink
dns records. It is a list of domain names for private hosted zones that
are to be shared by multiple clusters' privatelink records. They must
exist in the same account as the Hive cluster and be associated with the
privatelink VPC. When a cluster's domain is a subdomain of any item in
the list, the privatelink record will be created in the zone of the item
that most closely matches. Otherwise a new private hosted zone will be
created for the cluster.