Fix flaky TestGetName
There was a bug when shortName happened to be generated as a string
of length >= 246 as the test would expect that shortName + "-deploy"
did not exceed 253 characters and was not truncated. However, the
test previously randomly generated a shortName of length [1, 253),
which causes this test to flake and fail occasionally.

This commit creates a shortBase that at maximum is the longest
possible base that has room for a hash so that it will never be
truncated.

Signed-off-by: Michael Shen <mshen@redhat.com>