ClusterPool: hold until all clusters deleted, subject to maxConcurrent
Previously, deleting a ClusterPool would cause hive to issue Delete()s
for all unassigned clusters and then remove the primary finalizer from
the ClusterPool, allowing it to be garbage collected. The result was
that those unassigned clusters, as well as any claimed clusters, would
be "orphaned" until they were cleaned up.

With this commit, we still issue those Delete()s for unassigned clusters
(claimed clusters are still not touched). But we retain the finalizer,
causing the ClusterPool to remain in existence, until all of the pool's
clusters -- claimed and unclaimed -- are gone.

During this process, we maintain a new status condition of Type
`DeletionPossible`, whose Message includes a list of CDs pending
cleanup. A couple of things to note about this condition:
- It is unset (Status `Unknown`) until the ClusterPool is `Delete()`d.
- You will _usually_ not see the Status go to `"True"` because the
finalizer will be removed very quickly, causing the ClusterPool to
disappear (unless there are other finalizers present).

This commit also resolves a long-standing TODO: we now honor the
ClusterPool's `MaxConcurrent` setting when issuing the `Delete()`s
during this cleanup process. This also takes into account any clusters
which were already installing or deleting prior to the ClusterPool being
deleted.

This commit also resolves a latent bug whereby deleting a ClusterClaim
after a ClusterPool was deleted would fail to delete the corresponding
ClusterDeployment. Now it does, and such CDs are correctly counted
against `MaxConcurrent` along with the unassigned ones.

HIVE-1557