hiveutil aws-tag-deprovision: don't require a cluster
Changes made via #1874 / 0f4b1de caused hiveutil deprovisioning commands
to try to load up a cloud credentials secret upon startup. When run in
standalone mode, this doesn't work (unless you have configured some very
specific environment variables) BUT the deprovisioner is able to work
without it as long as your cloud creds are in one of the usual expected
places (e.g. credentials file, env vars). This is in fact how those
deprovisioners have always worked when invoked standalone.

However, prior to this commit, if we're entirely unable to contact the
cluster API, we would abort the command. With this commit, we instead
print a warning and proceed.

Note that we're only doing this for aws-tag-deprovision at this time, as
it's the only deprovisioner used in standalone mode with any regularity
(to our knowledge).

HIVE-2142