AWS MachinePool: Allow public and/or private subnets
Previously, via HIVE-1224 / #1234 / a27ef2e, we started filtering out
public subnets if they were provided in an AWS MachinePool. This led to
two issues:
- Users could not create machines in public subnets at all (HIVE-2196).
- In PrivateLink, subnets are always private, even though they may
appear public based on their attributes (HIVE-1805).

With this commit, we relax the restrictions such that, if the number of
subnets matches the number of availability zones, we allow them whether
they're public or private.

Here's a complete description of the new behavior:
- We still accept an empty Subnets list, as before.
- If the number of subnets matches the number of availability zones, we
allow them, whether they're public or private. (We still validate that
we were given exactly one subnet per AZ.)
- If the number of subnets is double the number of availability zones,
we follow the existing behavior: filter out the ones we detect are
public and pass through only the private ones.
- Any other number of subnets is an error.

HIVE-1805
HIVE-2196