New install failure reason: NoWorkerNodesReady
Investigating AuthenticationOperatorDegraded, which is currently the
fifth-most common install failure, can be annoying because install
failures are not caused by the Authentication operator failing; instead
it is a symptom of another failure. This PR represents a repeatable
finding where a log line indicating that 0 worker nodes have joined the
cluster causes AuthenticationOperatorDegraded.

Returning this instead will point us in a better direction to start
investigating the install failure. If we find that it makes up all cases
of AuthenticationOperatorDegraded we can remove that failure reason.

This error code is very similar to AWSNoWorkerNodes and
GCPNoWorkerNodes, however based on our history neither of them have
triggered for the past 90 days (and anecdotally even longer than that).
I suspect the regex is no longer valid since they were originally
created and this new regex can be their replacement.

Signed-off-by: Michael Shen <mshen@redhat.com>

Co-authored-by: Eric Fried <efried@redhat.com>