Azure custom resource group support
Day 0, the installer knows how to accept a custom resource group from
the install-config and use it for the cluster's cloud resources.
However, hive previously had no visibility to that field (in the
install-config or elsewhere); so for day 2 operations including
hibernation and additional MachinePools as well as for deprovisioning,
we were using the default (<cluster-infra-id>-rg). For machine pools,
this resulted in failures in the operation. In hibernation and
deprovisioning, the operation would "succeed" but no-op. For hibernation
this meant that machines were still running (incurring the associated
costs) even though the cluster claimed it was hibernating. For
deprovisioning, this meant the cloud resources were leaked.

This commit adds a new ClusterDeployment.Spec.ClusterMetadata section
for `Azure` containing a `ResourceGroupName` field, set as follows:
1. When provisioning, we discover it from the installer metadata during
installation.
2. When retrofitting a ClusterDeployment that was adopted or created
before custom Azure resource group support, we first attempt to discover
that resource group via the Infrastructure object in the remote cluster.
3. If that fails, we try parsing the install-config referenced by
ClusterDeployment.Spec.Provisioning.InstallConfigSecretRef.

Note that we will assign the default value if we're using the installer
metadata or the install-config and it's unset there. In any other case
where we can't figure it out (e.g. we can't find the infrastructure
object; or we can't parse the install-config) we just don't set it. The
consequence of that is...

This field is now required for any of the operations that use it -- if
it is absent, those operations will fail. While we don't expect this
condition to persist, as a fallback the user can manually set the value
if necessary.

HIVE-2204