Ensure hive-trusted-cabundle earlier
In HIVE-2198 / #1999 / 253408ed8 we started injecting a ConfigMap
containing a CA bundle for better proxy support. Alas, we were only
injecting it in the provisioning path. For hives that never had the
earlier code, this probably would have been fine, as the ConfigMap would
continue to exist until the cluster was deleted, whereupon it could
properly be mounted on the deprovisioning pods. But for spokes that
existed prior to upgrading to that code, and which were never
re-reconciled before they were deleted (or, conceivably, if a user
deleted the ConfigMap at some point in there) the ConfigMap would not
exist, resulting in mount failures attempting to bring up the
deprovisioning pods.

This commit moves the ConfigMap injection earlier in the flow, ensuring
it exists for both provisioning and deprovisioning paths.

HIVE-2207