ClusterProvision.Spec.Metadata => .MetadataJSON
ClusterProvision.Spec.Metadata is of type runtime.RawExtension. It
doesn't work (maybe it never did) -- we think because a
runtime.RawExtension can only hold a (marshaled) runtime.Object, and the
installer's ClusterMetadata type isn't one.

This commit creates a new field, MetadataJSON, of type `[]byte`, where
we store the ClusterMetadata as a raw JSON string. All uses of the old
Metadata field are replaced with MetadataJSON.

Needed by HIVE-2204