Move OpenShift discovery earlier, and cache it
We support running hive on non-OpenShift clusters, but we do a couple of
things differently if that's the case. Previously we were dynamically
discovering OpenShift-ness in the places we needed to know. With this
commit, we discover it at init time, and cache the result in the
reconcile object so we can consult it later.

Needed for HIVE-2210

(cherry picked from commit bf502af9e5b4911b188f85213997b3bceebac616)